Create mp_Program
So this is hopefully the beginning of a multi-purpose program for myself. 
It'll ask my name and the what I'd like to do.
Whether it's add, subtract, multiply, or divide.
With more features planned.
I'm not sure of the limitations of CS50 sandbox but I'd like to create functions to create a modular program. (If that makes sense)